checkpoint

+++ src/src_dotnet/JAStudio.Core/Note/Collection/CardStudyingStatus.cs
@@ -2,12 +2,12 @@ namespace JAStudio.Core.Note.Collection;
 
 public class CardStudyingStatus
 {
-    public int NoteId { get; }
+    public long NoteId { get; }
     public string CardType { get; }
     public bool IsSuspended { get; }
     public string NoteTypeName { get; }
 
-    public CardStudyingStatus(int noteId, string cardType, bool isSuspended, string noteTypeName)
+    public CardStudyingStatus(long noteId, string cardType, bool isSuspended, string noteTypeName)
     {
         NoteId = noteId;
         CardType = cardType;
